+++ test/OpenTelemetry.Exporter.Zipkin.Tests/ZipkinExporterTests.cs
@@ -92,7 +92,7 @@ namespace OpenTelemetry.Exporter.Zipkin.Tests
         }
 
         [Fact]
-        public void SuppresssesInstrumentation()
+        public void SuppressesInstrumentation()
         {
             const string ActivitySourceName = "zipkin.test";
             Guid requestId = Guid.NewGuid();