+++ src/OpenTelemetry.Exporter.OpenTelemetryProtocol/CHANGELOG.md
@@ -219,7 +219,7 @@ Released 2021-Feb-04
 
 Released 2021-Jan-29
 
-* Changed `OltpTraceExporter` class and constructor from internal to public.
+* Changed `OtlpTraceExporter` class and constructor from internal to public.
   ([#1612](https://github.com/open-telemetry/opentelemetry-dotnet/issues/1612))
 
 * In `OtlpExporterOptions.cs`: Exporter options now include a switch for Batch